make_job in lm7K/TestHomeDimerAtom/
make_job --bg
in TEST.bg/ TEST.atom.bg. (Main job file is jobque).

make_job --bg --continue is to generate job files which restart
calculations (check the end of save file). This is not tested yet. Be
careful.

job16core is useful to run TEST.pjsub/jobque.
It picks devide jobque+"number id" files into 16 jobs each, and
submit jobs.

+++ lm7K/TestHomoDimerAtom/jobquescript.py
@@ -55,6 +55,20 @@ source extra.bash
 REPLACEHERE___
 """
 
+jobtemppjsubnrel = """\
+#!/bin/bash
+#PJM -L "node=1"
+#PJM -L "rscgrp=fx-single"
+#PJM -L "elapse=48:00"
+#PJM -X
+#PJM --no-stging
+PATH=${HOME}/bin:${HOME}/local/bin:$PATH
+source atomlist.bash
+source homodimerdistance.bash
+#source extra.bash
+source extra_nrel.bash
+REPLACEHERE___
+"""
 
 jobtempbgnrel = """\
 #!/bin/bash
@@ -102,15 +116,37 @@ REPLACEHERE___
 """
 
 #open(sys.argv[2],'rt').read() #open('jobtemplate','rt').read()
-if sys.argv[2]=='--bg':
+print sys.argv
+if '--bg' in sys.argv:
     mode='bg'
-    print 'jobque file for backgound mode'
-elif sys.argv[2]=='--pjsub':
-    print 'jobque file for pjsub mode'
+    print ##### Generate jobque files for backgound mode #####'
+elif '--pjsub' in sys.argv:
+    print '##### Genarate jobque files for pjsub mode #####'
     mode='pjsub'
 else:
     print 'no option for jobquescript'
     sys.exit()
+
+if '--atom' in sys.argv:
+    print 'Atom mode'
+    atom=1
+else :
+    print 'Dimer mode'
+    atom=0
+
+if '--nrel' in sys.argv:
+    print 'nrel mode'
+    nrel=1
+    if(mode=='bg'): mode='bgnrel'
+    if(mode=='pjsub'): mode='pjsubnrel'
+else :
+    print 'rel mode'
+    nrel=0
+if '--continue' in sys.argv:
+    cont=1
+else:
+    cont=0
+
 
 patt    = string.split( open(sys.argv[1],'rt').read(),'\n') 
 for comm in patt:
@@ -119,29 +155,33 @@ for comm in patt:
 # distance
     dat = keydata(comm, 'distance=',';')
     distance = rangereal(float(dat[0]),float(dat[1]),0.1)
+#    print distance
+    if(atom==1): distance=['0.0']
     fsmoms = keydata(comm, 'fsmom=',',')
     rstars = keydata(comm, 'rstar=',',')
     enddat = enddata(comm)
     print 
     print '=== ',distance,' ',fsmoms,' ',rstars,' ',enddat
     print comm
-    
-    jobque     = open('jobque','a')
+    jjj='jobque'
+    if cont==1: jjj=jjj+'.continue'
+    jobque     = open(jjj,'a')
     #jobquebgnrel = open('jobque.bgnrel','a')
     #jobqueqsub   = open('jobque.qsub','a')
     for dis in distance:
         for fsmom in fsmoms:
             for rstar in rstars:
-                command='jobmoldimer1.py '+dis+', fsmom='+fsmom+'@ rstar='+rstar+'@ '+enddat
+                if(atom==1): command='jobatom1.py '+dis+', fsmom='+fsmom+'@ rstar='+rstar+'@ '+enddat
+                if(atom==0): command='jobmoldimer1.py '+dis+', fsmom='+fsmom+'@ rstar='+rstar+'@ '+enddat
 
                 # Generate ctrl files
-                jobfile = open('jobtempexec','w')
-                fff=string.replace(jobtemp,'REPLACEHERE___',command)
-                jobfile.write(fff)
-                jobfile.close()
-                os.system("bash jobtempexec")
-                os.system("rm jobtempexec")
-
+                if cont==0:
+                    jobfile = open('jobtempexec','w')
+                    fff=string.replace(jobtemp,'REPLACEHERE___',command)
+                    jobfile.write(fff)
+                    jobfile.close()
+                    os.system("bash jobtempexec")
+                    os.system("rm jobtempexec")
                 # directory name for working (where we have ctrl files).
                 f=open('ctrldir','r')
                 ddd=f.read()
@@ -149,22 +189,42 @@ for comm in patt:
                 #print 'cwd    = ',os.getcwd()
                 f.close()
 
+                # Check previous calculations. tail save.dimer, and findout where we restart.
+                if cont==1:
+                    initic=-1
+                    tailsave=commands.getstatusoutput('tail -1 '+ddd+'/save.dimer')
+                    if tailsave[0]==0: 
+                        print 'tail savedimer=',tailsave[1]
+                        pwe=tailsave[1].split('pwemax=')[1].split(' ')[0]
+                        bzw=tailsave[1].split('bzw=')[1].split(' ')[0]
+                        print 'pwe bzw from the end of save.dimer=',pwe,bzw
+                        if(pwe==2 and bzw==.01):
+                            initic=1
+                        else:
+                            initic=int(pwe)
+                else:
+                    initic=-9999
+
                 # all ctrl files sorted.
                 clines=''
                 ctrlfiles=string.split(commands.getoutput('ls -1 '+ddd+'/ctrl.dimer.*'),'\n')
                 for ic in sorted(ctrlfiles):
-                    cname= 'ctrl.dimer'+string.split(ic,'ctrl.dimer')[1]
-                    clines= clines+'cp '+cname+' ctrl.dimer;bash ctrl.dimer\n'
+                    pwexx=string.split(ic,'ctrl.dimer.')[1]
+                    cname= 'ctrl.dimer.'+pwexx
+                    if int(pwexx) >= initic:
+                        clines= clines+'cp '+cname+' ctrl.dimer;bash ctrl.dimer; rm mixm.dimer\n'
                 #print clines
-                   
+
                 # job file
                 jobname=string.replace(string.replace(command,' ','_'),'$','')
+                if cont==1: jobname=jobname+'.continue'
                 print '  jobfile=',jobname
                 jobfile = open(jobname,'w')
-                if(mode=='bg'): jobtempmmm=jobtempbg
+                if(mode=='bg'):    jobtempmmm=jobtempbg
                 if(mode=='pjsub'): jobtempmmm=jobtemppjsub
+                if(mode=='bgnrel'):jobtempmmm=jobtempbgnrel
+                if(mode=='pjsubnrel'):jobtempmmm=jobtemppjsubnrel
                 fff=string.replace(jobtempmmm,'REPLACEHERE___','cd '+ddd+'\n'+clines)
-                #fff=string.replace(jobtempbgnrel,'REPLACEHERE___','cd '+ddd+'\n'+clines)
                 jobfile.write(fff)
                 jobfile.close()
